Header, SQL, Config and welcome updates
Included new class, config, to store some site configuration values. It
is intended that the user will be able to change these via the admin
interface in later development.
SQL has basic authentication (to be changed) and a query method.
The header has some logic to display a site description if possible -
this is specified in the config.php file.