default.css - button implementation improvements
Switched from using a styled div to a styled a link for the buttons.
This allows the entire button to act as the link, rather than just the
centre text.
A new background gradient was added, with support included for webkit,
IE9+ and firefox 3.6+. On other browsers, the button will be displayed
without the gradient.
Removed pixels where possible, except for the text shadow- experimented
with using em (eg 0.1em), but the results were not very pretty.