Parent Authentication changes
The way parents are authenticated has been changed, in order to take
account of a scenario where a student could be their own parent (from
the perspective of the system, naturally) - this could feasibly be the
case for an older or mature student.
A flag is now set when logging in that indicates that the current user
has access to the parent functionality. This is what is checked when
the user panel determines what buttons to allow.