Deprecated adminPanel
With the development of the userPanel, the adminPanel is no longer
needed. Its functions have been incorporated into the general userPanel
and will only be made visible to logged in admins, and the index
redirector has been set up accordingly.
In other development, some new accessor methods have been added to
libraries/sql to deal with userid <-> username translation, which will
be necessary for the forthcoming 'compose message' feature.