Messaging - send message
Initial send message form has been created, and also deals with an
error caused by the user name being incorrect. Now need to create the
backend handler for the form.
Note: this also makes use of an HTML5 feature that allows fields to be
marked as required, this is quite handy in ensuring forms are validated
correctly before input.