Messaging - Initial send message functionality done
Sending messages can now be done from within StrathTutor.
A new backend library, message_new, does all the processing from the
input form. It uses the existing libraries to check that the given
usernames exist, escapes all input for security, and sends the message.
In the event the given username doesn't exist, it will redirect to the
previous page with an error message.