Basic student lesson view!
A basic lession view has been constructed for students to view their
lessons, and with a little modification this can also be used for
tutors. Currently displays as a table, although this will change to a
calendar later in development.
Also fixed a whole host of problems with the admin lessons panel,
including some fairly dodgy table code that didn't close half the tags.