Lessons - removed endTime, added duration
Major change to lessons- it made more sense to include standard lesson
times than a seperate end time. As a result, the code has been
refactored to make use of this, as well as the corresponding database
changes.
Drop down options in user_booking.php are populated directly from the
database.
The database also now holds a cost multiplier for each lesson duration,
in preparation for upcoming invoicing features.