File management - part 2
The main processing of files should now be in place, with the ability
to upload files from the individual lesson view for tutors/admins only.
The backend file management script will check the file to ensure it is
the correct size and type, and that it does not already exist, and
should handle setting success/error messages and redirection.
A file view now needs to be implemented for individual lessons and this
feature can be tested.