A little tidying - part 3
Renamed user lessons page to be more consistent with other features.
Refactored references to this page as needed.
Also added a title to user lesson booking page, again to be more
consistent with other pages.
Also deprecated user_report_display: it's not needed as the basic
lesson view has all the required information.